Improve autosave and backup feature (#2194)

+++ src/main/java/net/sf/jabref/logic/autosaveandbackup/BackupManager.java
@@ -32,6 +32,7 @@ import org.apache.commons.logging.LogFactory;
 /**
  * Backups the given bib database file from {@link BibDatabaseContext} on every {@link BibDatabaseContextChangedEvent}.
  * An intelligent {@link ExecutorService} with a {@link BlockingQueue} prevents a high load while making backups and rejects all redundant backup tasks.
+ * This class does not manage the .bak file which is created when opening a database.
  */
 public class BackupManager {
 
@@ -74,6 +75,10 @@ public class BackupManager {
 
     @Subscribe
     public synchronized void listen(@SuppressWarnings("unused") BibDatabaseContextChangedEvent event) {
+        startBackupTask();
+    }
+
+    private void startBackupTask() {
         try {
             executor.submit(backupTask);
         } catch (RejectedExecutionException e) {
@@ -90,7 +95,9 @@ public class BackupManager {
         bibDatabaseContext.getMetaData().unregisterListener(this);
         executor.shutdown();
         try {
-            Files.delete(backupPath);
+            if (Files.exists(backupPath) && !Files.isDirectory(backupPath)) {
+                Files.delete(backupPath);
+            }
         } catch (IOException e) {
             LOGGER.error("Error while deleting the backup file.", e);
         }
@@ -114,6 +121,7 @@ public class BackupManager {
         if (originalFile.isPresent()) {
             backupManager.originalPath = originalFile.get().toPath();
             backupManager.backupPath = getBackupPath(backupManager.originalPath);
+            backupManager.startBackupTask();
             bibDatabaseContext.getDatabase().registerListener(backupManager);
             bibDatabaseContext.getMetaData().registerListener(backupManager);
             runningInstances.add(backupManager);
